Refactor critical pairs/sequences
Move logic of aggregating all pairs/sequences
for a grammar out of the Analysis modules and
into the CLI.

This belongs in the CLI because it involves
strategies for calculating things in parallel,
which is not the role of the Analysis modules.

Furthermore, when we use an arbitrary Monad
for each category, there will be no universal way
to calculate things in parallel for any monad.